apps/shared/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "shared"
version = "0.1.0"
description = "Shared library for crosshot-ai agents"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "redis>=7.1.0",
    "openai>=2.17.0",
    "pyyaml>=6.0.3",
]
